+++ spec/gherkin/c_lexer_spec.rb
@@ -15,7 +15,7 @@ module Gherkin
       it_should_behave_like "a Gherkin lexer lexing tags"
       it_should_behave_like "a Gherkin lexer lexing doc_strings"
       it_should_behave_like "a Gherkin lexer lexing rows"
-      it_should_behave_like "parsing windows files"
+      it_should_behave_like "encoding"
     end
   end
 end