+++ spec/gherkin/native_lexer_spec.rb
@@ -23,7 +23,7 @@ module Gherkin
       it_should_behave_like "a Gherkin lexer lexing tags"
       it_should_behave_like "a Gherkin lexer lexing doc_strings"
       it_should_behave_like "a Gherkin lexer lexing rows"
-      it_should_behave_like "parsing windows files" unless ENV['GHERKIN_JS_NATIVE']
+      it_should_behave_like "encoding" unless ENV['GHERKIN_JS_NATIVE']
     end
   end
 end